summaryChat/requirements.txt
streamlit
google.generativeai
python-dotenv
youtube-transcript-api
pathlib